1) allow python inline expression in ${} for destination path, 2) added atac, chip example def JSON, 3) separated argparse from main module, 4) update DOC
